+++ wallets/get_received_chart.js
@@ -3,11 +3,11 @@ const asyncDetect = require('async/detect');
 const asyncFilterLimit = require('async/filterLimit');
 const asyncMap = require('async/map');
 const {formatTokens} = require('ln-sync');
+const {getAllInvoices} = require('ln-sync');
 const {getPayment} = require('ln-service');
 const moment = require('moment');
 const {returnResult} = require('asyncjs-util');
 
-const getAllInvoices = require('./get_all_invoices');
 const {segmentMeasure} = require('./../display');
 const {sumsForSegment} = require('./../display');