+++ wallets/index.js
@@ -1,6 +1,5 @@
-const getAllInvoices = require('./get_all_invoices');
 const getReceivedChart = require('./get_received_chart');
 const getReport = require('./get_report');
 const unlockWallet = require('./unlock_wallet');
 
-module.exports = {getAllInvoices, getReceivedChart, getReport, unlockWallet};
+module.exports = {getReceivedChart, getReport, unlockWallet};